feat: space game movements

+++ bundle/default/include/system/CameraSystem.hpp
@@ -4,6 +4,6 @@
 
 class CameraSystem : public System
 {
-  public:
-    void update(double deltaTime) override;
+public:
+  void update(float deltaTime) override;
 };
